==> Bulkseq/02_SalmonQuant_to_txi.Rmd <==
---
title: "02_SalmonQuant_to_txi"
author: "Joe and Luda"
date: "4/15/2022"
output: html_document
---

# this scipt takes results of salmon quantification and create gene by sample table

##Library
```{r}
library(tximport)
library(readr)
library(readxl)
```

##Load Data
```{r}
wd ='SG_Novogene_12192022'

# path to folder with the resulta of salmon run
dataFolder = 'SG_Novogene_12192022/Salmon_Quants'

#sample annotation
sampAnnotFile = 'SG_Novogene_12192022/SG_Novogene_Annot.csv'
# sample annotation file with sample names in the first column that can be used as file prefix
sampAnnot = read.csv(sampAnnotFile)

sampNames = sampAnnot$FileName

# paths to quantification files generated by salmon
files <- file.path(dataFolder, sampNames, "quant.sf")
all(file.exists(files))
# add sample names that will be used in txi object
names(files) = sampNames

```


# import transcript-level data, not summarized by gene
```{r}
txi <- tximport(files, type = "salmon", txOut=TRUE)
```

#===========
# create mapping between transcripts and genes to summarize by gene
#===========

# read one file from salmon to extract gene symbols from transcript IDs
```{r}
tab = read.table(file.path(dataFolder, sampNames[1], "quant.sf"), sep = '\t',header=T,row.names=1, stringsAsFactors = F)

##read in tgMap from refgenie
map = read.delim("SG_Novogene_12192022/tgMap.tsv", header=FALSE)
rownames(map)=map$V1

##Subset map to include transcripts only from tab
map <- map[rownames(tab),]
colnames(map) <- c("GeneID","TXName")

table(map[,2])
```

# run tximport with gene mapping -- All Files
```{r}
#TXI with all genes
txi <- tximport(files, type = "salmon", tx2gene = map)
names(txi)
head(txi$counts)

##Save output
saveRDS(txi, file = paste0(wd,'/Data/txi_data_all.rds'))
```
